requirements.txt
streamlit
pandas
openai
python-telegram-bot
